+++ aas-lib/src/main/java/de/fraunhofer/iosb/aas/lib/spi/AasDataAddress.java
@@ -21,8 +21,10 @@ import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 import de.fraunhofer.iosb.aas.lib.model.AasProvider;
+import org.eclipse.digitaltwin.aas4j.v3.model.KeyTypes;
 import org.eclipse.digitaltwin.aas4j.v3.model.Reference;
 import org.eclipse.digitaltwin.aas4j.v3.model.impl.DefaultReference;
+import org.eclipse.edc.connector.dataplane.http.spi.HttpDataAddress;
 import org.eclipse.edc.spi.EdcException;
 import org.eclipse.edc.spi.result.Result;
 import org.eclipse.edc.spi.types.domain.DataAddress;
@@ -31,6 +33,7 @@ import org.jetbrains.annotations.Nullable;
 import java.net.URL;
 import java.util.Base64;
 import java.util.HashMap;
+import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
@@ -38,9 +41,9 @@ import java.util.Optional;
 import static de.fraunhofer.iosb.aas.lib.model.impl.Service.CONCEPT_DESCRIPTIONS_PATH;
 import static de.fraunhofer.iosb.aas.lib.model.impl.Service.SHELLS_PATH;
 import static de.fraunhofer.iosb.aas.lib.model.impl.Service.SUBMODELS_PATH;
-import static org.eclipse.edc.spi.constants.CoreConstants.EDC_NAMESPACE;
 import static java.util.Collections.emptyMap;
 import static java.util.stream.Collectors.toMap;
+import static org.eclipse.edc.spi.constants.CoreConstants.EDC_NAMESPACE;
 
 /**
  * Inspired by org.eclipse.edc.connector.dataplane.http.spi.HttpDataAddress
@@ -51,9 +54,8 @@ import static java.util.stream.Collectors.toMap;
 public class AasDataAddress extends DataAddress {
 
     public static final String AAS_DATA_TYPE = "AasData";
-
+    public static final String OPERATION_NAME = KeyTypes.OPERATION.name().toLowerCase(Locale.ROOT);
     // See aas4j operation
-    public static final String OPERATION = "operation";
     private static final String ADDITIONAL_HEADER = "header:";
     private static final String METHOD = "method";
     private static final String PROVIDER = "AAS-Provider";
@@ -66,11 +68,11 @@ public class AasDataAddress extends DataAddress {
     }
 
     public boolean isOperation() {
-        return this.hasProperty(OPERATION);
+        return this.hasProperty(OPERATION_NAME);
     }
 
     public @Nullable String getOperation() {
-        return isOperation() ? this.getStringProperty(OPERATION) : null;
+        return isOperation() ? this.getStringProperty(OPERATION_NAME) : null;
     }
 
     @JsonIgnore
@@ -87,7 +89,7 @@ public class AasDataAddress extends DataAddress {
 
     private AasProvider getProvider() {
         Object provider = Optional
-        .ofNullable(super.getProperties().get(EDC_NAMESPACE + PROVIDER))
+                .ofNullable(super.getProperties().get(EDC_NAMESPACE + PROVIDER))
                 .orElse(super.getProperties().get(PROVIDER));
         if (provider instanceof AasProvider) {
             return (AasProvider) provider;
@@ -130,18 +132,18 @@ public class AasDataAddress extends DataAddress {
 
     private String referenceChainAsPath() {
         StringBuilder urlBuilder = new StringBuilder();
-
         for (var key : getReferenceChain().getKeys()) {
             var value = key.getValue();
             String[] toAppend = switch (key.getType()) {
-                case ASSET_ADMINISTRATION_SHELL -> new String[] { SHELLS_PATH, b64(value) };
-                case SUBMODEL -> new String[] { SUBMODELS_PATH, b64(value) };
-                case CONCEPT_DESCRIPTION -> new String[] { CONCEPT_DESCRIPTIONS_PATH, b64(value) };
-                case SUBMODEL_ELEMENT, SUBMODEL_ELEMENT_COLLECTION, SUBMODEL_ELEMENT_LIST ->
-                    new String[] { urlBuilder.indexOf("/submodel-elements/") == -1 ? "/submodel-elements/".concat(value)
-                            : ".".concat(value) };
-                default ->
-                    throw new EdcException(new IllegalStateException("Element type not recognized: %s".formatted(key)));
+                case ASSET_ADMINISTRATION_SHELL -> new String[]{ SHELLS_PATH, b64(value) };
+                case SUBMODEL -> new String[]{ SUBMODELS_PATH, b64(value) };
+                case CONCEPT_DESCRIPTION -> new String[]{ CONCEPT_DESCRIPTIONS_PATH, b64(value) };
+                case SUBMODEL_ELEMENT, SUBMODEL_ELEMENT_COLLECTION, SUBMODEL_ELEMENT_LIST, PROPERTY,
+                     ANNOTATED_RELATIONSHIP_ELEMENT, RELATIONSHIP_ELEMENT, DATA_ELEMENT, MULTI_LANGUAGE_PROPERTY, RANGE, FILE, BLOB,
+                     REFERENCE_ELEMENT, CAPABILITY, ENTITY, EVENT_ELEMENT, BASIC_EVENT_ELEMENT, OPERATION ->
+                        new String[]{ urlBuilder.indexOf("/submodel-elements/") == -1 ? "/submodel-elements/".concat(value)
+                                : ".".concat(value) };
+                default -> throw new EdcException(new IllegalStateException("Element type not recognized: %s".formatted(key)));
             };
 
             urlBuilder.append(String.join("/", toAppend));
@@ -156,7 +158,7 @@ public class AasDataAddress extends DataAddress {
         return Base64.getEncoder().encodeToString(toBeEncoded.getBytes());
     }
 
-    private Reference getReferenceChain() {
+    public Reference getReferenceChain() {
         var referenceChain = properties.get(REFERENCE_CHAIN);
 
         if (referenceChain == null) {
@@ -168,6 +170,18 @@ public class AasDataAddress extends DataAddress {
         }
 
         throw new EdcException(new IllegalStateException(("Faulty reference chain: %s").formatted(referenceChain)));
+    }
+
+    public HttpDataAddress asHttpDataAddress() {
+        HttpDataAddress.Builder httpDataAddress = HttpDataAddress.Builder.newInstance();
+        this.getProvider().getHeaders().forEach(httpDataAddress::addAdditionalHeader);
+
+        return httpDataAddress
+                .baseUrl(this.getAccessUrl().getContent().toString())
+                .method(this.getMethod())
+                .path(this.getPath())
+                .build();
+
     }
 
     @JsonPOJOBuilder(withPrefix = "")
@@ -204,7 +218,7 @@ public class AasDataAddress extends DataAddress {
          * consumer to provider (during "compaction" phase when serializing the DA)
          */
         public Builder operation(String operation) {
-            this.property(OPERATION, operation);
+            this.property(OPERATION_NAME, operation);
             return this;
         }