+++ aas-lib/src/testFixtures/java/de/fraunhofer/iosb/aas/test/StringMethods.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.fraunhofer.iosb.app.testutils;
+package de.fraunhofer.iosb.aas.test;
 
 import org.eclipse.digitaltwin.aas4j.v3.dataformat.core.SerializationException;
 import org.eclipse.digitaltwin.aas4j.v3.dataformat.json.JsonSerializer;