+++ edc-extension4aas/src/main/java/de/fraunhofer/iosb/app/aas/mapper/Mapper.java
@@ -16,10 +16,12 @@
 package de.fraunhofer.iosb.app.aas.mapper;
 
 import de.fraunhofer.iosb.aas.lib.model.AasProvider;
-import org.eclipse.digitaltwin.aas4j.v3.model.Referable;
+import org.eclipse.digitaltwin.aas4j.v3.model.Identifiable;
 import org.eclipse.edc.connector.controlplane.asset.spi.domain.Asset;
 
-public interface Mapper<T extends Referable> {
+import java.util.function.BiFunction;
 
-    Asset map(T element, AasProvider provider);
+public interface Mapper<T extends Identifiable> extends BiFunction<T, AasProvider, Asset> {
+
+    Asset apply(T element, AasProvider provider);
 }