+++ edc-extension4aas/src/main/java/de/fraunhofer/iosb/app/edc/CleanUpService.java
@@ -50,8 +50,8 @@ public class CleanUpService implements SelfDescriptionChangeListener {
 
     @Override
     public void removed(Service service) {
-        if (service.environment() != null) {
-            servicePipeline.execute(service.environment());
+        if (service.getEnvironment() != null) {
+            servicePipeline.execute(service.getEnvironment());
         }
     }
 
@@ -59,7 +59,7 @@ public class CleanUpService implements SelfDescriptionChangeListener {
     public void removed(Registry registry) {
         Optional.ofNullable(registry.services())
                 .orElse(List.of()).stream()
-                .map(Service::environment)
+                .map(Service::getEnvironment)
                 .forEach(servicePipeline::execute);
     }
 
@@ -126,7 +126,7 @@ public class CleanUpService implements SelfDescriptionChangeListener {
                     .step(new ContractRegistrar(
                             Objects.requireNonNull(contractDefinitionStore),
                             Objects.requireNonNull(policyDefinitionStore),
-                            monitor))
+                            monitor, "dummy-participant-id"))
                     .monitor(monitor)
                     .build();